Removed InsertionRequests/TerminationRequests. Replaced by NotificationRequests
