+++ src/test/resources/journeys.conf
@@ -15,20 +15,17 @@
 # Configure here your journeys. A journey is a sequence of requests at a certain load.
 
 journeys {
-
-  notification-journey = {
-    description = "notification journey"
+  universal-credit-liability-notification-journey = {
+    description = "Universal Credit Liability Notification journey"
 
     load = 4
     feeder = data/data.csv
 
     parts = [
-      sendNotification
+      send-universal-credit-liability-notifications
     ]
   }
-
 }
-
 
 # Default behaviour is to run all journeys. If that is not what you need you can specify the list of journeys to run
 #journeysToRun = [
