--- src/test/scala/uk/gov/hmrc/universalcreditliabilityapi/InsertionRequests.scala
@@ -1,61 +1,0 @@
-/*
- * Copyright 2025 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.universalcreditliabilityapi
-
-import io.gatling.core.Predef._
-import io.gatling.http.Predef._
-import io.gatling.http.request.builder.HttpRequestBuilder
-import uk.gov.hmrc.performance.conf.ServicesConfiguration
-
-import java.util.UUID
-
-object InsertionRequests extends ServicesConfiguration {
-
-  def nationalInsuranceNumber: String = RandomNino.next()
-  val authToken: String               = AuthHelper.getAuthToken
-
-  val baseUrl: String       = baseUrlFor("universal-credit-notification")
-  val route: String         = "/notification"
-  val originatorId: String  = ""
-  val correlationId: String = UUID.randomUUID().toString
-
-  def notificationBody: String =
-    s"""
-    |{
-    |  "nationalInsuranceNumber": "$nationalInsuranceNumber",
-    |  "universalCreditRecordType": "{{universalCreditRecordType}}",
-    |  "universalCreditAction": "{{universalCreditAction}}",
-    |{% if universalCreditAction == "Insert" %}  "dateOfBirth": "2002-10-10",\n{% endif %}
-    |{% if universalCreditAction == "Terminate" %}  "liabilityEndDate": "2025-08-19",\n{% endif %}
-    |  "liabilityStartDate": "2025-08-19"
-    |}
-    |""".stripMargin
-
-  val sendNotification: HttpRequestBuilder =
-    http("Send insertion notification")
-      .post(s"$baseUrl$route")
-      .headers(
-        Map(
-          "authorization"        -> authToken,
-          "content-type"         -> "application/json",
-          "gov-uk-originator-id" -> originatorId,
-          "correlationId"        -> correlationId
-        )
-      )
-      .body(PebbleStringBody(notificationBody))
-      .check(status.is(204))
-}
